refactor(codezap): global 패키지 하위 항목을 용도에 따라 패키징
