+++ backend/src/main/java/codezap/global/auditing/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package codezap.global.domain;
+package codezap.global.auditing;
 
 import java.time.LocalDateTime;
