+++ backend/src/main/java/codezap/global/cors/WebCorsConfiguration.java
@@ -1,4 +1,4 @@
-package codezap.global;
+package codezap.global.cors;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
