+++ backend/src/main/java/codezap/global/swagger/SpringDocConfiguration.java
@@ -1,4 +1,4 @@
-package codezap.swagger;
+package codezap.global.swagger;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
