+++ backend/src/main/java/codezap/template/domain/Template.java
@@ -7,7 +7,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
-import codezap.global.domain.BaseTimeEntity;
+import codezap.global.auditing.BaseTimeEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
